requirements.txt
beautifulsoup4
opensecrets-crpapi
tqdm